build should only use app as cb if cb is undefined
build(app, cb) was defined to allow app to be optional, but it should
only shift the arguments if cb is undefined not simply if app is a function
(which if passed in  it is always a fn).

By adding the cb is undefined check we can support the following build signatures:

 - build()
 - build(app, cb)
 - build(cb)
 - build(app, null)

Added tests to describe and verify these extra 3 scenarios.